+++ src/globaltoken/hardfork.h
@@ -11,8 +11,8 @@ namespace Consensus {
     struct Params;
 };
 
-arith_uint256 GetAlgoPowLimit(int algo, const Consensus::Params& consensusParams);
-arith_uint256 GetAlgoPowLimit(int algo);
+arith_uint256 GetAlgoPowLimit(uint8_t algo, const Consensus::Params& consensusParams);
+arith_uint256 GetAlgoPowLimit(uint8_t algo);
 bool IsHardForkActivated(uint32_t blocktime, const Consensus::Params& consensusParams);
 bool IsHardForkActivated(uint32_t blocktime);
 int64_t GetPoWTargetTimeSpan(uint32_t blocktime, const Consensus::Params& consensusParams);
