+++ src/validation.h
@@ -468,7 +468,7 @@ extern VersionBitsCache versionbitscache;
  */
 int32_t ComputeBlockVersion(const CBlockIndex* pindexPrev, const Consensus::Params& params);
 
-int32_t SetAlgo(int algo);
+int32_t SetAlgo(uint8_t algo);
 
 /** Reject codes greater or equal to this can be returned by AcceptToMemPool
  * for transactions, to signal internal conditions. They cannot and should not
